requirements.txt
flask==3.1.0
gunicorn==23.0.0
flask-limiter==3.7.0
bleach==6.1.0

# Google Cloud Services
google-generativeai==0.8.3
google-cloud-translate==3.19.0
google-cloud-aiplatform==1.60.0
firebase-admin==6.5.0

# Testing & Code Quality
pytest==8.0.0
pytest-cov==4.1.0
pytest-mock==3.14.0

# Additional utilities
python-dotenv==1.0.0
